# Reworked encode_signed_integer() and decode_signed_integer() implementations for much better performance