Up to 80% improved performance of Integer Encode:

%%timeit
encode_signed_integer(-1946141750)
2.44 µs ± 201 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)

%%timeit
encode_signed_integer_new(-1946141750)
468 ns ± 138 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)


Up to 60% improved performance of Integer Decode:

%%timeit
decode_int_old(b'\x8c\x00\x3b\xca')
644 ns ± 48.2 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)

%%timeit
decode_int_new(b'\x8c\x00\x3b\xca')
264 ns ± 57.1 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)