@Finndersen 2021-11-23
@eerimoq any idea whats up? The changes I made should have no effect on the failing test cases for codec parsing etc and they all pass on my system