@eerimoq 2021-11-23
The problem is that Github Actions installs pyparsing 3.x.y, which is incompatible with asn1tools. All tests pass with pyparsing 2.m.n. I guess we should update asn1tools to be compatible with the latest version of pyparsing.