@eerimoq 2021-11-28
asn1tools should work with both old and new pyparsing now.